Config

+++ src/pages/bookshelf/events/book-card-events.ts
@@ -1,8 +1,4 @@
-import {
-   changeBookCategory,
-   deleteBook,
-   renameBook,
-} from '../../../database/book-repository.ts';
+import { changeBookCategory, deleteBook, renameBook } from '../../../database/book-repository.ts';
 import { getCategoryByName } from '../../../database/category-repository.ts';
 
 import renderBookshelf from '../features/render-bookshelf.ts';
@@ -22,9 +18,7 @@ export default function bindBookCardEvents() {
 
       if (clickedButton.matches('.rename-book-btn')) {
          const title = await openFormFor('book-name') as string;
-         if (title) {
-            await renameBook(bookId, title);
-         }
+         if (title) await renameBook(bookId, title);
       } else if (clickedButton.matches('.change-book-category-btn')) {
          const selectedCategory = await openFormFor(
             'category-selection',
@@ -35,9 +29,7 @@ export default function bindBookCardEvents() {
          }
       } else if (clickedButton.matches('.delete-book-btn')) {
          const confirmation = await openFormFor('confirmation') as boolean;
-         if (confirmation) {
-            await deleteBook(bookId);
-         }
+         if (confirmation) await deleteBook(bookId);
       }
 
       await renderBookshelf();
